Refactor(endpoint): add models
